sanphamct

+++ src/main/java/com/example/duantotnghiepgiaythethaonova/dto/HinhAnhDTO.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class HinhAnhDTO {
 	
-	private Long hinhAnhid;
+	private Integer hinhAnhid;
 	
 	private String tenAnh;
